+++ custom_components/maintenance_supporter/frontend/strategy/chunks/chunk-TBQ5N7W5.js
@@ -1,4 +1,4 @@
-import{a as r}from"./chunk-XJVTMRPN.js";var a=r`
+import{a as r}from"./chunk-RS3O3VYV.js";var a=r`
   ha-card { overflow: hidden; }
   .card-content {
     padding: 16px;